Revised the 'maxage' argument of the ClarkLDF and ClarkCapeCod functions so that it no longer served two purposes. Now it only pertains to the maximum age to which losses are projected.
The second purpose is accommodated by the new argument adol.age, which is the age of the average date of loss of occurrences within an origin period.

Another new argument, origin.width, specifies the width of an origin period in the same units as the ages of the triangle. This is particularly useful when the triangle's ages are not uniformly spaces, so that the default calculation of origin.width = mean(diff(ages)) might be misleading.

Of course, the new arguments are only pertinent when adol=TRUE, i.e., when the ages of the evaluations are considered to be the time from the average date of loss of the origin period.

RUnit tests were added and the documentation was updated.

git-svn-id: https://chainladder.googlecode.com/svn/trunk@273 e5f87290-e53f-0410-917e-5545ed4be688