Changed 1st argument of ClarkLDF and ClarkCapeCod from 'data' to 'Triangle' to emphasize that the data must be in the form of a matrix-shaped actuarial "triangle" and to be more consistent with other functions in ChainLadder, e.g., MackChainLadder. Help files were updated.
Explicitly set the 'factr' control option for the "L-BFGS-B" method of 'optim' with the result that convergence is now more consistent between R's 32-bit and 64-bit environments.

Set the lower limit of 'U' (LDF method) and 'ELR' (Cape Cod) to sqrt(.Machine$double.eps). This fixed a glitch that often resulted in infeasible initial solutions. These infeasible solutions were discovered with the 'trace' control option while debugging the convergence issue. Optim had been correcting for the infeasibility automatically, which is why the situation had gone undetected.

The unit test of the qincurred run was corrected.

git-svn-id: https://chainladder.googlecode.com/svn/trunk@274 e5f87290-e53f-0410-917e-5545ed4be688