Fixes for tail extrapolation in Vignette and Chainladder
1) The calculation for the tail log-linear extrapolation given in the
vignette had a minor error.  This has been corrected, and the result now
agrees with the results of MackChainLadder(RAA, tail=TRUE).
2) The calculation of the tail using the log-linear extrapolation in
ChainLadder.R had a potential error - when clratios has values of less
than unity they are dropped, but the extrapolation was started from a
quantity indexed by the length of f, not the value of fn.  This changes
the results if clratios has a pattern like e.,g.: ... 1.1,
0.98,1.01,0.005 (i.e. a link ratio less than unity which is not the last
value)
3) Minor fix to the comments in ChainLadder.R and MackChainLadder.R,
fixing notation for alpha which is now consistent with the documentation
and Mack's original paper.