Small fixes in the code
Some fixing in the documentation, removing the french in the coments and the code.
There were also some typo in the docs and in the examples.
We also ran styler and improved the documentation.
We added copyright info to the code.
We removed flaky documentation that was not allowing r cmd check to pass.